requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.1
future==0.18.2
idna==2.10
Jinja2==3.0.1
joblib==1.0.1
livereload==2.6.3
lunr==0.5.8
lxml==4.6.3
Markdown==3.3.2
MarkupSafe==2.0.1
mkdocs==1.1.2
mkdocs-material==7.1.5
mkdocs-material-extensions==1.0.1
nltk==3.6.2
psycopg2==2.8.6
Pygments==2.9.0
pymdown-extensions==8.2
PySocks==1.7.1
pytz==2020.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-futures==1.0.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
stem==1.8.0
tornado==6.1
torrequest==0.1.0
tqdm==4.61.0
urllib3==1.26.6
whitenoise==5.2.0